requirements.txt
torch
torchvision
timm
numpy
scikit-learn
matplotlib
seaborn
fpdf
tqdm
pyyaml